add typescript user preferences to the schema
The schema is based on the latest TS version so it might not apply
when using older TS versions from the project or might get outdated
if we forget to update it for new versions of typescript.

Also, enabled "allowRenameOfImportPath" and "includeCompletionsForImportStatements"
in default settings. Those were not enabled by default by the server.

The "includeCompletionsForModuleExports" and "includeCompletionsWithInsertText"
are enabled by default in the server but added to default settings anyway
for transparency.