Implement a custom "find callers" command (#90)
typescript-language-server implements a custom "textDocument/calls"
request which is a lot like normal "find references" but it skips all
references that are not calling the symbol. So it filters a lot of
noise like references to function being imported or just passed around.

The "textDocument/calls" request supports finding both callers
(implemented here) and a sort of an opposite mode of operation where it
finds all functions called inside the selected symbol (function). I find
that kinda useless so didn't implement it but it can be easily added
by creating another `lsp_typescript_calls` command and passing
"direction": "outgoing" argument to it.